Add files via upload
its a Java program that simulates an exit poll for the UK election in July 2024. This program will collect the number of constituencies won by each party and determine if any party has won the majority in the House of Commons (326 constituencies).
This program uses the Scanner class to read the number of constituencies won by each party from the user. It then checks if any party has won the majority (326 constituencies) and prints the result. You can run this program in any Java IDE or compile it using a command-line tool.